Refactor DefaultStateRegistry
