+++ server/src/main/java/io/spine/server/model/Model.java
@@ -24,7 +24,9 @@ import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
+import com.google.protobuf.Message;
 import io.spine.core.CommandClass;
+import io.spine.protobuf.Messages;
 import io.spine.server.aggregate.Aggregate;
 import io.spine.server.aggregate.AggregateClass;
 import io.spine.server.aggregate.AggregatePart;
@@ -245,6 +247,43 @@ public class Model {
             classes.put(nameOf(cls), modelClass);
         }
         return (EntityClass<?>) modelClass;
+    }
+
+    public Message getDefaultState(Class<? extends Entity> cls) {
+        final DefaultStateRegistry registry = DefaultStateRegistry.getInstance();
+        final Message state = createDefaultState(cls);
+        @SuppressWarnings("unchecked")
+        // cast is safe because this type of messages is saved to the map
+        final Message defaultState = registry.putOrGet(cls, state);
+        return defaultState;
+    }
+
+    private Message createDefaultState(Class<? extends Entity> cls) {
+        final Class<? extends Message> stateClass = getStateClass(cls);
+        final Message result = Messages.newInstance(stateClass);
+        return result;
+    }
+
+    /**
+     * Obtains the class of the entity state.
+     */
+    private Class<? extends Message> getStateClass(Class<? extends Entity> cls) {
+        //TODO:2017-08-21:alexander.yevsyukov: Use the below code to overcome the issue in the test
+        // environment entity, which is created with the generic type <I> variable, which is
+        // resolved only at the level of enclosed test suite (RecordStorageShould).
+        // Since there is no way to get the value of the type variable (which we need to construct
+        // model entity class), we use this type name checking and return state class via static
+        // method call. The rest of the classes should follow general declaration contract.
+        //
+        // Once these tests are reworked, eliminate the code in the if() statement below, and inline
+        // the EntityClass.getStateClass() method.
+        //
+        if (cls.getName().contains("TestCounterEntity")) {
+            return EntityClass.getStateClass(cls);
+        }
+
+        Class<? extends Message> result = asEntityClass(cls).getStateClass();
+        return result;
     }
 
     /**
